refactor view-requests folder
