+++ src/pages/dashboard/requests/view-requests/admin/index.tsx
@@ -1,14 +1,14 @@
 import Cancel from "@/components/custom/Icons/Cancel";
 import { SheetClose, SheetContent } from "@/components/ui/sheet";
 import { useMediaQuery } from "react-responsive";
-import Summary from "./Summary";
+import Summary from "../Summary";
 import { Dialog, DialogClose, DialogContent, DialogTrigger } from "@/components/ui/dialog";
 import { useState } from "react";
-import AssignReviewer from "../components/AssignReviewer";
+import AssignReviewer from "../../components/AssignReviewer";
 import { useLocation } from "react-router-dom";
 import ArrowRight from "@/components/custom/Icons/ArrowRight";
 // import useUserStore from "@/store/user-store";
-import WriteReview from "../components/WriteReview";
+import WriteReview from "../../components/WriteReview";
 import { useRequestsStore } from "@/store/RequestFormStore";
 import Loader from "@/components/custom/Loader";
 import { useRole } from "@/hooks/useRole";
